+++ logging/all.go
@@ -1,18 +1,18 @@
 package logging
 
 import (
-	"log"
-	"net/http"
-	"os"
-	"time"
-
 	"github.com/BSick7/go-api/intercept"
+	"log/slog"
+	"net/http"
+	"time"
 )
 
-func LogAllRequests(prefix string) intercept.OnResponseFunc {
-	stdNoTime := log.New(os.Stderr, prefix, 0)
-
+func LogAllRequests(logger *slog.Logger) intercept.OnResponseFunc {
 	return func(r *http.Request, data intercept.ResponseData, duration time.Duration) {
-		stdNoTime.Printf("%s %d %s %s %s", duration, data.StatusCode(), r.Method, r.RequestURI, data.Body())
+		logger.InfoContext(r.Context(), data.Body(),
+			slog.Duration("duration", duration),
+			slog.Int("status_code", data.StatusCode()),
+			slog.String("method", r.Method),
+			slog.String("request_uri", r.RequestURI))
 	}
 }